requirements.txt
ipython==7.18.1
Pillow==8.0.1
reportlab==3.5.55
